+++ src/Presentation/SmartStore.Web/Administration/Controllers/ShoppingCartController.cs
@@ -6,9 +6,7 @@ using SmartStore.Core.Domain.Orders;
 using SmartStore.Services.Catalog;
 using SmartStore.Services.Customers;
 using SmartStore.Services.Helpers;
-using SmartStore.Services.Localization;
 using SmartStore.Services.Security;
-using SmartStore.Services.Stores;
 using SmartStore.Services.Tax;
 using SmartStore.Web.Framework.Controllers;
 using SmartStore.Web.Framework.Security;
@@ -19,45 +17,32 @@ namespace SmartStore.Admin.Controllers
     [AdminAuthorize]
     public class ShoppingCartController : AdminControllerBase
     {
-        #region Fields
-
         private readonly ICustomerService _customerService;
         private readonly IDateTimeHelper _dateTimeHelper;
         private readonly IPriceFormatter _priceFormatter;
-		private readonly IStoreService _storeService;
         private readonly ITaxService _taxService;
         private readonly IPriceCalculationService _priceCalculationService;
-        private readonly IPermissionService _permissionService;
-        private readonly ILocalizationService _localizationService;
-        #endregion
 
-        #region Constructors
-
-        public ShoppingCartController(ICustomerService customerService,
-            IDateTimeHelper dateTimeHelper, IPriceFormatter priceFormatter,
-			IStoreService storeService, ITaxService taxService,
-			IPriceCalculationService priceCalculationService,
-            IPermissionService permissionService, ILocalizationService localizationService)
+        public ShoppingCartController(
+            ICustomerService customerService,
+            IDateTimeHelper dateTimeHelper, 
+            IPriceFormatter priceFormatter,
+            ITaxService taxService,
+			IPriceCalculationService priceCalculationService)
         {
-            this._customerService = customerService;
-            this._dateTimeHelper = dateTimeHelper;
-            this._priceFormatter = priceFormatter;
-			this._storeService = storeService;
-            this._taxService = taxService;
-            this._priceCalculationService = priceCalculationService;
-            this._permissionService = permissionService;
-            this._localizationService = localizationService;
+            _customerService = customerService;
+            _dateTimeHelper = dateTimeHelper;
+            _priceFormatter = priceFormatter;
+            _taxService = taxService;
+            _priceCalculationService = priceCalculationService;
         }
-
-        #endregion
         
-        #region Methods
-
-        //shopping carts
         public ActionResult CurrentCarts()
         {
-            if (!_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
+            {
                 return AccessDeniedView();
+            }
 
             return View();
         }
@@ -67,7 +52,7 @@ namespace SmartStore.Admin.Controllers
         {
 			var gridModel = new GridModel<ShoppingCartModel>();
 
-			if (_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+			if (Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
 			{
 				var q = new CustomerSearchQuery
 				{
@@ -77,14 +62,15 @@ namespace SmartStore.Admin.Controllers
 					PageSize = command.PageSize
 				};
 
-				var customers = _customerService.SearchCustomers(q);
+                var guestStr = T("Admin.Customers.Guest").Text;
+                var customers = _customerService.SearchCustomers(q);
 				
 				gridModel.Data = customers.Select(x =>
 				{
 					return new ShoppingCartModel
 					{
 						CustomerId = x.Id,
-						CustomerEmail = x.IsGuest() ? T("Admin.Customers.Guest").Text : x.Email,
+						CustomerEmail = x.IsGuest() ? guestStr : x.Email,
 						TotalItems = x.CountProductsInCart(ShoppingCartType.ShoppingCart)
 					};
 				});
@@ -109,15 +95,15 @@ namespace SmartStore.Admin.Controllers
         {
 			var gridModel = new GridModel<ShoppingCartItemModel>();
 
-			if (_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+			if (Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
 			{
-				var customer = _customerService.GetCustomerById(customerId);
+                decimal taxRate;
+                var customer = _customerService.GetCustomerById(customerId);
 				var cart = customer.GetCartItems(ShoppingCartType.ShoppingCart);
 
 				gridModel.Data = cart.Select(sci =>
 				{
-					decimal taxRate;
-					var store = _storeService.GetStoreById(sci.Item.StoreId);
+					var store = Services.StoreService.GetStoreById(sci.Item.StoreId);
 
 					var sciModel = new ShoppingCartItemModel
 					{
@@ -126,12 +112,13 @@ namespace SmartStore.Admin.Controllers
 						ProductId = sci.Item.ProductId,
 						Quantity = sci.Item.Quantity,
 						ProductName = sci.Item.Product.Name,
-						ProductTypeName = sci.Item.Product.GetProductTypeLabel(_localizationService),
+						ProductTypeName = sci.Item.Product.GetProductTypeLabel(Services.Localization),
 						ProductTypeLabelHint = sci.Item.Product.ProductTypeLabelHint,
 						UnitPrice = _priceFormatter.FormatPrice(_taxService.GetProductPrice(sci.Item.Product, _priceCalculationService.GetUnitPrice(sci, true), out taxRate)),
 						Total = _priceFormatter.FormatPrice(_taxService.GetProductPrice(sci.Item.Product, _priceCalculationService.GetSubTotal(sci, true), out taxRate)),
 						UpdatedOn = _dateTimeHelper.ConvertToUserTime(sci.Item.UpdatedOnUtc, DateTimeKind.Utc)
 					};
+
 					return sciModel;
 				});
 
@@ -150,12 +137,12 @@ namespace SmartStore.Admin.Controllers
             };
         }
 
-
-        //wishlists
         public ActionResult CurrentWishlists()
         {
-            if (!_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+            if (!Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
+            {
                 return AccessDeniedView();
+            }
 
             return View();
         }
@@ -165,7 +152,7 @@ namespace SmartStore.Admin.Controllers
         {
 			var gridModel = new GridModel<ShoppingCartModel>();
 
-			if (_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+			if (Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
 			{
 				var q = new CustomerSearchQuery
 				{
@@ -175,14 +162,15 @@ namespace SmartStore.Admin.Controllers
 					PageSize = command.PageSize
 				};
 
-				var customers = _customerService.SearchCustomers(q);
+                var guestStr = T("Admin.Customers.Guest").Text;
+                var customers = _customerService.SearchCustomers(q);
 
 				gridModel.Data = customers.Select(x =>
 				{
 					return new ShoppingCartModel
 					{
 						CustomerId = x.Id,
-						CustomerEmail = x.IsGuest() ? T("Admin.Customers.Guest").Text : x.Email,
+						CustomerEmail = x.IsGuest() ? guestStr : x.Email,
 						TotalItems = x.CountProductsInCart(ShoppingCartType.Wishlist)
 					};
 				});
@@ -207,15 +195,15 @@ namespace SmartStore.Admin.Controllers
         {
 			var gridModel = new GridModel<ShoppingCartItemModel>();
 
-			if (_permissionService.Authorize(StandardPermissionProvider.ManageOrders))
+			if (Services.Permissions.Authorize(StandardPermissionProvider.ManageOrders))
 			{
-				var customer = _customerService.GetCustomerById(customerId);
+                decimal taxRate;
+                var customer = _customerService.GetCustomerById(customerId);
 				var cart = customer.GetCartItems(ShoppingCartType.Wishlist);
 
 				gridModel.Data = cart.Select(sci =>
 				{
-					decimal taxRate;
-					var store = _storeService.GetStoreById(sci.Item.StoreId);
+					var store = Services.StoreService.GetStoreById(sci.Item.StoreId);
 
 					var sciModel = new ShoppingCartItemModel
 					{
@@ -224,12 +212,13 @@ namespace SmartStore.Admin.Controllers
 						ProductId = sci.Item.ProductId,
 						Quantity = sci.Item.Quantity,
 						ProductName = sci.Item.Product.Name,
-						ProductTypeName = sci.Item.Product.GetProductTypeLabel(_localizationService),
+						ProductTypeName = sci.Item.Product.GetProductTypeLabel(Services.Localization),
 						ProductTypeLabelHint = sci.Item.Product.ProductTypeLabelHint,
 						UnitPrice = _priceFormatter.FormatPrice(_taxService.GetProductPrice(sci.Item.Product, _priceCalculationService.GetUnitPrice(sci, true), out taxRate)),
 						Total = _priceFormatter.FormatPrice(_taxService.GetProductPrice(sci.Item.Product, _priceCalculationService.GetSubTotal(sci, true), out taxRate)),
 						UpdatedOn = _dateTimeHelper.ConvertToUserTime(sci.Item.UpdatedOnUtc, DateTimeKind.Utc)
 					};
+
 					return sciModel;
 				});
 
@@ -247,7 +236,5 @@ namespace SmartStore.Admin.Controllers
                 Data = gridModel
             };
         }
-
-        #endregion
     }
 }