+++ src/Presentation/SmartStore.Web/Models/Boards/ForumPostModel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using SmartStore.Web.Framework.Modelling;
+using SmartStore.Web.Models.Customer;
 
 namespace SmartStore.Web.Models.Boards
 {
@@ -15,7 +16,6 @@ namespace SmartStore.Web.Models.Boards
         
         public int CustomerId { get; set; }
         public bool AllowViewingProfiles { get; set; }
-        public string CustomerAvatarUrl { get; set; }
         public string CustomerName { get; set; }
         public bool IsCustomerForumModerator { get; set; }
         public bool IsCustomerGuest { get; set; }
@@ -38,5 +38,6 @@ namespace SmartStore.Web.Models.Boards
 
         public int CurrentTopicPage { get; set; }
 
+        public CustomerAvatarModel Avatar { get; set; }
     }
 }